requirements.txt
fastapi
uvicorn[standard]
aioredis>=2.0.0